requirements-updated.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
ffmpeg-python==0.2.0
flatbuffers==1.12
fonttools==4.39.3
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
iopath==0.1.9
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mtcnn==0.1.1
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
portalocker==2.7.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tomli==2.0.1
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
yacs==0.1.8
zipp==3.15.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@ebb26b5b0d914a9fc196da303e536892165a7113